IA-725 Fix updating schedules to properly recurse child tasks
Before calling the readyToSave method of the task controller, the task path is copied. The copy does not include pointers to objects that are used to run the task and **only** includes properties used to archive the task result.

Thus, the child path elements are `nil` for a copied path since these are used to carry **state** information rather than intended to carry results, themselves.