Belt and suspenders for de-duping scheduled activities
Fix for BRIDGE-2350

Before sending an updated list of the schedules, check that the instances included in the array are unique and only include if this is the first instance. Using the first instance because the code needs to recursively search through the task results for an associated schedule, in case the schedule that runs the task is not the same and the schedule that needs to be marked as finished.