Split BridgeApp framework into smaller logical groupings
BridgeApp is growning to include a lot of “shared” utilities that are only used in
one or two applications. Additionally, it includes stuff that keeps the base files
from conforming to App Extension requirements. Because of this, I have broken
it into three frameworks:

BridgeApp : Cross-platform and app extension connection to BridgeSDK.

BridgeAppUI: Convenience view controllers and app delegate for use with
an iOS application.

DataTracking: Support for medication, triggers, and symptom tracking.