Fix related date range handling issues
Bridge appears to treat the startDate as non-inclusive. Also, when we convert the singleton report date to a date-only and then extract the date components for year/month/day, because DateComponentsFormatter always uses the local time zone we’re ending up with 2000-12-31 instead of 2001-01-01, but this will vary by timezone. Since we’ve already saved reports this way that the app will still need to be able to retrieve in the future, “fixing” that without workarounds to accommodate it would break the app. This commit includes the workarounds which are needed regardless, but not an actual “fix” for the underlying date issue.