Fix handling for the time zone
The server reports was always returning the timestamp date in GMT so the time zone wasn’t being saved. Additionally, for the case were the user is switching the time zone during the day and it is displayed to the user, the time zone abbreviation does not include the abbrevation that is expected for a US timezone (PDT, for example) and instead shows “GMT-7”. This fixes both of those issues.