Simplify convertion to/from answer format
I also commented out but did not delete a bunch of code that does not appear to be used and I refactored the names a bit for clarity.

Basically, a lot of the confusion comes from attempting to convert from Swift objects to Obj-c objects where SBBJSONValue conforms to JSON serializable NSObject classes and RSDJSONSerializable conforms to Decodable objects (which NSNumber is *not*). Fun, fun.