MP2-263 Add data group profile item type
As a part of adding the data group profile type, I removed the assertions about unrecoginized decodable types and replaced with a print statement. This is to allow additive profile item types to be reverse-compatible to older versions of an app.

Additionally, I also changed the protocol to use a class instead of a struct. This allows for a more appropriate setter that handles the state of the item. Alternatively, the `value { get set}` pattern could be replaced with a pattern that explicitly calls out how those values are updated. Structs, every time you mutate the struct by setting a property on it, it will copy the struct to a new object. The only reason this was working was because the getter wasn’t ever looking at a cached value and only looks to the backing store.