Copy the logging step before displaying to the participant
When displayed as part of a tracking task, the logging step is mutated to change the title and text that is displayed to the user. Because of this, when shown from the meds button, it was showing the wrong language. This is a class and *not* a struct, therefore it needs to be copied explicitly. Ooops.